move the staleness tests into the imager history suite
